==> ML-project-main/notebooks/01_inspect_resume_dataset.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns


# Load the CSV file (adjust path if needed)
df = pd.read_csv('../data/resumes_dataset.csv')

# View top 5 rows
df.head()


df.columns


df.info()


df.isnull().sum()


df['Category'].value_counts()


import nltk
nltk.download('punkt')
